SpectralMatching and uniform result path names
