+++ workflow/rules/common.smk
@@ -60,28 +60,28 @@ def get_final_output():
     # dictionary of rules and its output files
     rule_dict = {"fileconversion" : expand(os.path.join("data", "mzML", "{dataset}.mzML"), dataset=DATASET),
                 "preprocessing" : [expand(os.path.join("results", "Interim", "mzML", "PCpeak_{dataset}.mzML"), dataset=DATASET),
-        expand(os.path.join("results", "Interim", "Preprocessed", "FFM_{dataset}.featureXML"), dataset=DATASET),
-        expand(os.path.join("results", "Interim", "Preprocessed", "Filtered_{sample}.featureXML"), sample=SUBSAMPLES),
+        expand(os.path.join("results", "Interim", "Preprocessing", "FFM_{dataset}.featureXML"), dataset=DATASET),
+        expand(os.path.join("results", "Interim", "Preprocessing", "Filtered_{sample}.featureXML"), sample=SUBSAMPLES),
         expand(os.path.join("results", "Interim", "mzML", "PCfeature_{sample}.mzML"), sample=SUBSAMPLES),
-        expand([os.path.join("results", "Interim", "Preprocessed", "MapAligned_{sample}.featureXML"), os.path.join("results", "Interim", "Preprocessed", "MapAligned_{sample}.trafoXML")], sample=SUBSAMPLES),
+        expand([os.path.join("results", "Interim", "Preprocessing", "MapAligned_{sample}.featureXML"), os.path.join("results", "Interim", "Preprocessing", "MapAligned_{sample}.trafoXML")], sample=SUBSAMPLES),
         expand(os.path.join("results", "Interim", "mzML", "Aligned_{sample}.mzML"), sample=SUBSAMPLES),
-        expand([os.path.join("results", "Interim", "Preprocessed", "MFD_{sample}.featureXML"), os.path.join("results", "Interim", "Preprocessed", "MFD_{sample}.consensusXML")], sample=SUBSAMPLES),
-        expand(os.path.join("results", "Interim", "Preprocessed", "consenus_features_unfiltered.consensusXML")),
-        expand(os.path.join("results", "Interim", "Preprocessed", "consenus_features.consensusXML")),
-        expand(os.path.join("results", "Interim", "Preprocessed", "FeatureMatrix.tsv")),
-        expand(os.path.join("results", "Preprocessed", "FeatureTables", "FeatureMatrix_{sample}.tsv"), sample=SUBSAMPLES),
-        expand(os.path.join("results", "Preprocessed", "FeatureMatrix.tsv"))],
-                "requantification" : [expand([os.path.join("results", "Interim", "Requantified", "Complete.consensusXML"), os.path.join("results", "Interim", "Requantified", "Missing.consensusXML")]),
-        expand(os.path.join("results", "Interim", "Requantified", "Complete_{sample}.featureXML"), sample=SUBSAMPLES),
-        expand(os.path.join("results", "Interim", "Requantified", "MetaboliteNaN.tsv")),
-        expand(os.path.join("results", "Interim", "Requantified", "FFMID_{sample}.featureXML"), sample=SUBSAMPLES),
-        expand(os.path.join("results", "Interim", "Requantified", "Merged_{sample}.featureXML"), sample=SUBSAMPLES),
-        expand(os.path.join("results", "Interim", "Requantified", "MFD_{sample}.featureXML"), sample=SUBSAMPLES),
-        expand(os.path.join("results", "Interim", "Requantified", "IDMapper_{sample}.featureXML"), sample=SUBSAMPLES),
-        expand(os.path.join("results", "Interim", "Requantified", "consenus_features_unfiltered.consensusXML")),
-        expand(os.path.join("results", "Interim", "Requantified", "consenus_features.consensusXML")),
-        expand(os.path.join("results", "Interim", "Requantified", "FeatureMatrix.tsv")),
-        expand(os.path.join("results", "Requantified", "FeatureMatrix.tsv"))],
+        expand([os.path.join("results", "Interim", "Preprocessing", "MFD_{sample}.featureXML"), os.path.join("results", "Interim", "Preprocessing", "MFD_{sample}.consensusXML")], sample=SUBSAMPLES),
+        expand(os.path.join("results", "Interim", "Preprocessing", "consenus_features_unfiltered.consensusXML")),
+        expand(os.path.join("results", "Interim", "Preprocessing", "consenus_features.consensusXML")),
+        expand(os.path.join("results", "Interim", "Preprocessing", "FeatureMatrix.tsv")),
+        expand(os.path.join("results", "Preprocessing", "FeatureTables", "FeatureMatrix_{sample}.tsv"), sample=SUBSAMPLES),
+        expand(os.path.join("results", "Preprocessing", "FeatureMatrix.tsv"))],
+                "requantification" : [expand([os.path.join("results", "Interim", "Requantification", "Complete.consensusXML"), os.path.join("results", "Interim", "Requantification", "Missing.consensusXML")]),
+        expand(os.path.join("results", "Interim", "Requantification", "Complete_{sample}.featureXML"), sample=SUBSAMPLES),
+        expand(os.path.join("results", "Interim", "Requantification", "MetaboliteNaN.tsv")),
+        expand(os.path.join("results", "Interim", "Requantification", "FFMID_{sample}.featureXML"), sample=SUBSAMPLES),
+        expand(os.path.join("results", "Interim", "Requantification", "Merged_{sample}.featureXML"), sample=SUBSAMPLES),
+        expand(os.path.join("results", "Interim", "Requantification", "MFD_{sample}.featureXML"), sample=SUBSAMPLES),
+        expand(os.path.join("results", "Interim", "Requantification", "IDMapper_{sample}.featureXML"), sample=SUBSAMPLES),
+        expand(os.path.join("results", "Interim", "Requantification", "consenus_features_unfiltered.consensusXML")),
+        expand(os.path.join("results", "Interim", "Requantification", "consenus_features.consensusXML")),
+        expand(os.path.join("results", "Interim", "Requantification", "FeatureMatrix.tsv")),
+        expand(os.path.join("results", "Requantification", "FeatureMatrix.tsv"))],
                 "GNPSexport" : [expand(os.path.join("results", "Interim", "GNPSexport", "filtered.consensusXML")),
         expand(os.path.join("results", "GNPSexport", "MSMS.mgf")),
         expand(os.path.join("results", "GNPSexport", "FeatureQuantificationTable.txt")),
@@ -93,9 +93,10 @@ def get_final_output():
                 "sirius" : [expand(os.path.join("results", "Interim", "Sirius", "formulas_{sample}.mzTab"), sample=SUBSAMPLES),
         expand(os.path.join("results", "Sirius", "formulas_{sample}.tsv"), sample=SUBSAMPLES),
         expand(os.path.join("results", "annotations", "FeatureTable_sirius.tsv"))],
-                "spectralmatcher" : [expand(os.path.join("results", "Interim", "annotations", "MSMS.mzML")),
-        expand(os.path.join("results", "Interim", "annotations", "MSMSMatcher.mzTab")),
-        expand(os.path.join("results", "annotations", "FeatureTable_MSMS.tsv"))],
+                "spectralmatcher" : [expand(os.path.join("results", "Interim", "SpectralMatching", "MSMS.mzML")),
+        expand(os.path.join("results", "Interim", "SpectralMatching", "MSMSMatches.mzTab")),
+        expand(os.path.join("results", "Interim", "SpectralMatching", "FeatureMatrix.tsv")),
+        expand(os.path.join("results", "SpectralMatching", "FeatureMatrix.tsv"))],
                 "analogsearch" : [expand(os.path.join("results", "Interim", "annotations", "ms2query", "lib.txt")),
         expand(os.path.join("results", "GNPSexport", "results")),
         expand(os.path.join("results", "GNPSexport", "results", "MSMS.csv")),
